requirements.txt
streamlit~=1.42.0
google-generativeai
python-dotenv~=1.0.1
langchain~=0.3.18
PyPDF2~=3.0.1
chromadb
faiss-cpu
langchain_google_genai
protobuf~=5.29.3
langchain-google-genai~=2.0.9
langchain-community~=0.3.17